// src/api/generated/api.ts
export interface StaffDetailResponse {
    /**
     * 返回码，调用成功时返回 200
     */
    'code'?: string;
    /**
     * 调用失败时的错误信息
     */
    'message'?: string;
    'data'?: StaffDetailResponseData;
}
// ...
export interface StaffEnableResponse {
    /**
     * 响应码
     */
    'code': string;
    /**
     * 响应消息
     */
    'message': string;
    'data'?: StaffEnableResponseAllOfData;
}
// ...
export interface StaffApiInterface {
    /**
     * 禁用指定 ID 的员工账号
     * @summary 禁用员工
     * @param {string} id 员工 ID
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    disableStaff(id: string, options?: RawAxiosRequestConfig): AxiosPromise<StaffEnableResponse>;

    /**
     * 启用指定 ID 的员工账号
     * @summary 启用员工
     * @param {string} id 员工 ID
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    enableStaff(id: string, options?: RawAxiosRequestConfig): AxiosPromise<StaffEnableResponse>;

    /**
     * 查询员工详细信息（简版员工专用）。  **注意**：对于购买了核心人力的客户，请使用\"获取员工详细信息 MDD 接口\"。  id 和 code 至少传其一，用于唯一确定要查询的员工。 
     * @summary 查询员工详情
     * @param {string} [id] 业务数据 ID（id 和 code 至少传其一）
     * @param {string} [code] 业务数据人员 code（id 和 code 至少传其一）
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    queryStaff(id?: string, code?: string, options?: RawAxiosRequestConfig): AxiosPromise<StaffDetailResponse>;

}

/**
 * StaffApi - object-oriented interface
 */
export class StaffApi extends BaseAPI implements StaffApiInterface {
    /**
     * 禁用指定 ID 的员工账号
     * @summary 禁用员工
     * @param {string} id 员工 ID
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public disableStaff(id: string, options?: RawAxiosRequestConfig) {
        return StaffApiFp(this.configuration).disableStaff(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 启用指定 ID 的员工账号
     * @summary 启用员工
     * @param {string} id 员工 ID
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public enableStaff(id: string, options?: RawAxiosRequestConfig) {
        return StaffApiFp(this.configuration).enableStaff(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 查询员工详细信息（简版员工专用）。  **注意**：对于购买了核心人力的客户，请使用\"获取员工详细信息 MDD 接口\"。  id 和 code 至少传其一，用于唯一确定要查询的员工。 
     * @summary 查询员工详情
     * @param {string} [id] 业务数据 ID（id 和 code 至少传其一）
     * @param {string} [code] 业务数据人员 code（id 和 code 至少传其一）
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public queryStaff(id?: string, code?: string, options?: RawAxiosRequestConfig) {
        return StaffApiFp(this.configuration).queryStaff(id, code, options).then((request) => request(this.axios, this.basePath));
    }
}